Fix duplicate docstring in playwright control module
